Moving tensorFlow class into yise package for consistent organization of files.
